### 2017-08-01 #342 @SebastianoF
Hello!
@arokem @rougier @sealhuang @oesteban @effigies @nirum @ahwillia - you are all in neuroimage/neuroscience + Python - Is there anyone interested in peer-reviewing bruker2nifti as submitted for JOSS?
It is not fundamental, but it may help if you have worked with Bruker ParaVision software/formats before.

@neurolabusc @matthew-brett do you know someone potentially interested in peer-reviewing the code bruker2nifti?

Thanks!
