### 2017-08-01 #342 @effigies
I don't have any experience with Bruker, so if any other potential reviewers do, you should probably prioritize them over me. That said, as long as there are some good and bad Bruker files (for bad, ideally including an explanation of what's wrong, such as malformed header, truncated data segment, etc.) I think I could review this.
