### 2017-08-01 #342 @Kevin-Mattheus-Moerman
@mfroeling @jhuguetn (Jordi can you also ask Paul Groot) want to review this submission for [JOSS](joss.theoj.org)? Or do you know anybody that could help? If you accept we'll open up a review issue. The review process follows our [review guidelines](http://joss.theoj.org/about#reviewer_guidelines). Let me know if you have any questions.
