### 2017-08-01 #342 @Kevin-Mattheus-Moerman
@SebastianoF can you provide info in relation to @effigies 's question, i.e. do you have some tests/example Bruker files a reviewer could play with?
