### 2017-08-01 #342 @SebastianoF
@Kevin-Mattheus-Moerman There are bruker files a reviewer can play with, in the test_data folder.
They are a good starting point, and are the only one I could provide non-copyrighted at least at the moment.
This may make the test self-referential. Having a non-copyrighted external (and so unbiased) dataset is not that straightforward. I heard offline that @neurolabusc had in mind to create a benchmarking platform with a wider dataset available, where the tests will be unbiased. He already pointed out a subtle issue of frame inversion that happens in ParaVision version 5.1, raised in the issue [#11](https://github.com/SebastianoF/bruker2nifti/issues/11).
