### 2017-08-01 #342 @neurolabusc
A nice repository of images is here
  https://gitlab.com/naveau/bruker2nifti_qa/tree/master
I think a challenge is that while the Bruker file format is thoroughly documented, the files created by Bruker software often omit required fields from some of the files.
You are free to compare the results of this code versus my own
  https://github.com/neurolabusc/Bru2Nii
Since my code pre-dates this project (origins in 1990's) it tends to rely more on older Bruker files, and uses different heuristics to deal with missing fields. Of all my projects, this is the one I am least proud. Poorly conceived formats that are poorly executed necessarily lead to kludgey code. I think it is great that Sebastiano has created a clean-sheet implementation.
