### 2017-08-01 #342 @rougier
It's not exactly my expertise but I can (try to) help with review if needed.
