### 2017-08-01 #342 @Kevin-Mattheus-Moerman
@rougier Sure all help is welcome. I'm going to wait a couple of days to attract some more reviewers and then I'll open the formal review issue.
