### 2017-08-01 #342 @andrewjanke
I'm not entirely sure what is going on here, but I think I like what is being proposed. A code review for open journal publication of software?

I have a number of Bruker files here dating back to the early 90's, but as Chris points out the Bruker file format is rather like DICOM, it's a specification (effectively YAML but older) and how it has been used to determine orientation and other things from the acqup file is dependant on time, version, the author of the pulse sequence and a number of things. So 100% coverage is going to be hard.

Also realise that orientation is not guaranteed if the pulse sequence author hasn't done things "right", for example the phase read direction might be reversed and acqp will not record this.

And then there is PV6...

Short version: Let me know if I can help although not sure of entirely what is being asked just yet.
