### 2017-08-02 fc09d69
MAINT: add travis-ci testing
