### 2017-08-06 2bdf93f
DOC readme update
