### 2017-08-06 6f2aa73
WIP towards python2.7 python 3 backcompatibility
