+++ bruker2nifti/__init__.py
@@ -2,4 +2,4 @@ __author__     = "SebastianoF"
 __licence__    = "MIT"
 __repository__ = "https://github.com/SebastianoF/bruker2nifti"
 __all__ = ['_cores', '_getters', '_utils', 'converter', 'open_GUI']
-__version__ = '1.0.0'
+__version__ = '1.0.0a'
