### 2017-08-06 7c0ee99
ENH back compatibility test done - to be tested with travis-ci
