+++ bruker2nifti/parsers/bruker2nii.py
@@ -8,7 +8,7 @@ def main():
     $ python2 bruker2nifti_study -i input_file_path -o output_file_path
     """
 
-    parser = argparse.ArgumentParser(version=0.0)
+    parser = argparse.ArgumentParser()
 
     # pfo_study_bruker_input
     parser.add_argument('-i', '--input_study_folder',
