### 2017-08-07 ae3eca1
ENH travis ci email off. gitignore updated
