+++ .travis.yml
@@ -99,3 +99,6 @@ script:
 
 after_success:
     - if [ "${COVERAGE}" == "1" ]; then coveralls; fi
+
+notifications:
+  email: false
