### 2017-08-09 #342 @rougier
Ok. I can try to review but my review will be on a technical level mostly.
