### 2017-08-09 #342 @Kevin-Mattheus-Moerman
@rougier That is fine we can get started on the technical side first and I will recruit others to help with other aspects shortly
