### 2017-08-09 #354 @whedon
[REVIEW]: Bruker2nifti: Magnetic Resonance Images converter from Bruker ParaVision to Nifti format
**Submitting author:** @SebastianoF (<a href="http://orcid.org/0000-0001-9279-4765">Sebastiano Ferraris</a>)
**Repository:** <a href="https://github.com/SebastianoF/bruker2nifti" target ="_blank">https://github.com/SebastianoF/bruker2nifti</a>
**Version:** 1.0.0
