### 2017-08-09 #354 @Kevin-Mattheus-Moerman
@rougier you are officially assigned and can tick the boxes at the top of this review issue. However others are invited to review the same aspects and leave comments here as well. Thanks!
