### 2017-08-12 SebastianoF/bruker2nifti#17 @rougier
Dead links in the README
There are some dead (or faulty) links in the README:

* http://python-guide-pt-br.readthedocs.io/en/latest/dev/virtualenvs/
* https://github.com/SebastianoF/bruker2nifti/blob/master/bruker2nifti.rtfd.io

Also, is it "Mattew Brett" or "Matthew Brett" ?
