### 2017-08-12 SebastianoF/bruker2nifti#18 @rougier
Wiki page
The "Home" wiki page could be populated with the different links composing the wiki (they are on the right side but it might be more obvious if they were also on the "Home" page)
