### 2017-08-12 fb3bf32
DOC update after #17
