### 2017-08-12 SebastianoF/bruker2nifti#17 @SebastianoF
Typo and link corrected, thanks! Do you think the README is too pedantic? Maybe there is too much in there.
