### 2017-08-12 SebastianoF/bruker2nifti#18 @SebastianoF
Makes more sense. The link on the right side where always in random order...
