### 2017-08-12 SebastianoF/bruker2nifti#17 @rougier
No I think it's fine. What is not clear is what difference do you make with develop install and pyenv install. You have two lines with `pip install bruker2nifti` such that in the end it is not clear which one to use actually.

Also, you seem to imply that the only requirement is python 2.7, but you also need bumpy, no?
