### 2017-08-12 a16f0e1
DOC update after #17 and follow up
