### 2017-08-12 SebastianoF/bruker2nifti#17 @SebastianoF
I added some links to the requirements.txt file and to the development mode so to try to make things clear. The code actually works also with python 3. Is 2.7+ understood in this way?
