### 2017-08-12 37f8c7b
DOC update - travis ci badge added _ test
