### 2017-08-12 SebastianoF/bruker2nifti#17 @rougier
Not al all. For me, it was not working with Python 3. You should mention it explicitly.
