### 2017-08-12 SebastianoF/bruker2nifti#17 @SebastianoF
Yes, the released version is still only for Python 2.7. How to mange the discrepancy between the release version and the current version? Can I propose a new release on Pypi while under review?
