### 2017-08-12 SebastianoF/bruker2nifti#17 @SebastianoF
Maybe this is a question for the editor, will open a new issue for this!
