### 2017-08-12 SebastianoF/bruker2nifti#19 @SebastianoF
Releasing a new patch-version after JOSS submission
@Kevin-Mattheus-Moerman: quick question about JOSS policy on releasing new patch-versions while the review is in progress.

In the [issue #17](https://github.com/SebastianoF/bruker2nifti/issues/17) the reviewer noticed that the pip installable release is not compatible with python3 as stated in the wiki page and readme. This is correct as this new feature was added after the JOSS submission, and it is available in the current master branch on git, but it is not yet released on pypi yet.

Can I release a new patch-version (1.0.1) while the 1.0.0 is under submission, integrating the reviewer suggestions?
If yes, will the version at the review page [https://github.com/openjournals/joss-reviews/issues/354](https://github.com/openjournals/joss-reviews/issues/354) updated to the latest one?

Thanks!
