### 2017-08-13 7d44b61
ENH more test for converter and _cores.py. TYPOS in _cores
