+++ test/test_cores.py
@@ -1,10 +1,12 @@
 import os
 import numpy as np
+import warnings
+import sys
 
 from nose.tools import assert_equal
 from numpy.testing import assert_array_equal
 
-from bruker2nifti._cores import scan2struct
+from bruker2nifti._cores import scan2struct, write_struct
 
 here = os.path.abspath(os.path.dirname(__file__))
 root_dir = os.path.dirname(here)
@@ -12,11 +14,11 @@ root_dir = os.path.dirname(here)
 
 def test_scan2struct_with_banana_data_slope_no_slope():
 
-    pfo_study_in = os.path.join(root_dir, 'test_data', 'bru_banana', '1')
+    pfo_scan_in = os.path.join(root_dir, 'test_data', 'bru_banana', '1')
 
-    struct_no_slope_corrected = scan2struct(pfo_study_in, correct_slope=False)
+    struct_no_slope_corrected = scan2struct(pfo_scan_in, correct_slope=False)
 
-    struct_yes_slope_corrected = scan2struct(pfo_study_in, correct_slope=True)
+    struct_yes_slope_corrected = scan2struct(pfo_scan_in, correct_slope=True)
 
     parsed_slope_from_no = struct_no_slope_corrected['visu_pars_list'][0]['VisuCoreDataSlope']
     parsed_slope_from_yes = struct_yes_slope_corrected['visu_pars_list'][0]['VisuCoreDataSlope']
@@ -37,4 +39,30 @@ def test_scan2struct_with_banana_data_slope_no_slope():
                                          struct_yes_slope_corrected['nib_scans_list'][0].get_data(), decimal=4)
 
 
-test_scan2struct_with_banana_data_slope_no_slope()
+def test_write_struct():
+
+    with warnings.catch_warnings(record=True) as w:
+        warnings.simplefilter("always")
+        # Trigger a warning
+        pfo_study_in = os.path.join(root_dir, 'test_data', 'bru_banana')
+        scan2struct(pfo_study_in)
+        # manage warning in python 2 or python 3:
+        if sys.version_info[0] == 3:
+            assert len(w) == 1
+            assert issubclass(w[-1].category, UserWarning)
+        else:
+            import exceptions
+            assert len(w) == 1
+            assert issubclass(w[-1].category, exceptions.UserWarning)
+
+    pfo_scan_in = os.path.join(root_dir, 'test_data', 'bru_banana', '1')
+    banana_struct = scan2struct(pfo_scan_in)
+
+    pfo_output = os.path.join(root_dir, 'test_data', 'nifti_banana', 'write_struct_here')
+    if not os.path.exists(pfo_output):
+        os.system('mkdir {}'.format(pfo_output))
+
+    write_struct(banana_struct, pfo_output, fin_scan='test')
+
+    assert os.path.exists(os.path.join(pfo_output, 'acquisition_method.txt'))
+    assert os.path.exists(os.path.join(pfo_output, 'test.nii.gz'))
