### 2017-08-13 670fe1b
BUG versionig python bytecode string into utf-8 string due to Travis failure
