### 2017-08-13 7b7b41d
BUG wheel instead of sdist in yml due to versioning git check_output cwd problem
