+++ bruker2nifti/__init__.py
@@ -1,3 +1,4 @@
+import os
 from subprocess import check_output
 
 
@@ -6,9 +7,12 @@ __licence__    = "MIT"
 __repository__ = "https://github.com/SebastianoF/bruker2nifti"
 __all__ = ['_cores', '_getters', '_utils', 'converter', 'open_GUI']
 
+here = os.path.abspath(os.path.dirname(__file__))
+git_dir = os.path.dirname(here)
+
 # Describe the version relative to last tag
 command_git = ['git', 'describe', '--match', 'v[0-9]*']
-version_buf = check_output(command_git).rstrip()
+version_buf = check_output(command_git, cwd=git_dir).rstrip()
 
 # Exclude the 'v' for PEP440 conformity, see
 # https://www.python.org/dev/peps/pep-0440/#public-version-identifiers
