### 2017-08-13 5d06c1c
WIP coverall test 3: COVERAGE variable omitted
