### 2017-08-13 76ff34b
WIP coverall test 4 coverall restored for the latest wheel installation version
