### 2017-08-13 92c1fd3
WIP coverall COVERAGE restored and copy corrected
