### 2017-08-13 d4484c7
WIP coverall WIP nosetests outside the newly created folder in script
