### 2017-08-13 ad77ca8
WIP coverall no COVERAGE variable and COVERAGE env
