### 2017-08-13 d6035a6
ENH travis.yml restored to the originally pushed
