### 2017-08-13 a7e85e4
WIP travis + coverage. COVERAGE variable creating problems
