### 2017-08-13 a22a911
WIP while learning coverall.io coverage on external file redirected
