+++ nose_coverage.txt
@@ -1,0 +1,13 @@
+.......................
+Name                        Stmts   Miss  Cover
+-----------------------------------------------
+bruker2nifti/_cores.py        174     66    62%
+bruker2nifti/_getters.py      105     56    47%
+bruker2nifti/_utils.py        250     60    76%
+bruker2nifti/converter.py      68     15    78%
+-----------------------------------------------
+TOTAL                         597    197    67%
+----------------------------------------------------------------------
+Ran 23 tests in 0.556s
+
+OK
