### 2017-08-14 SebastianoF/bruker2nifti#20 @rougier
Instructions in the README
I'm testing installation and there's a small typo:

`cd bruker2nifti-master` -> `cd bruker2nifti`

For testing, you indicate `nose` but you do not explain how to use it and what should be the expected results (are there skipped tests, known errors, etc.)

For installation instructions, it is not clear that you provide alternative instruction for installation. Because of the structure of the README, it's like every step I necessary. Maybe you can have a main "installation" header and sub-headers for each installation (with a small introduction explaining these are alternative ways for installation).

For the non programming method, if would be good to provide the direct link to download the last release. For non programmer, it might be difficult to navigate the GitHub interface to find the right file. Also, the zip archive of the repo and the zip of the release are not really the same, you may prefer to use the "stable" release only since the current repo might be unstable.

For the initial description, could you add link to "Brukert ParaVision" and "nifty" ? It might be good also to explain things a bit more (what is the target audience, why it is necessary, real for example, etc.)
