### 2017-08-14 SebastianoF/bruker2nifti#21 @rougier
GUI
Is it possible to have tooltip explaining what are the different checkboxes or is it too much work?
This might be the limit of my expertise for the review but I found the GUI a bit "raw". A simple "help" button explaining things could help.
