### 2017-08-14 SebastianoF/bruker2nifti#22 @rougier
Documentation
The `index.rst` could benefit from an introduction (re-using the README) explaining the goal of the library, how to use, pointers to the GitHub repository, external links, etc).

A usage section using real-data would be also welcome.
