### 2017-08-14 #354 @rougier
I've started the review and reported minor issues on the project repository. I think the submission is in good shape and I'm now waiting for author corrections before finishing the review.
