### 2017-08-14 SebastianoF/bruker2nifti#23 @rougier
Paper
Can you add the DOI  for the two references in the bibliography?

Also, "Mattew Brett" -> "Matthew Brett"
