### 2017-08-14 SebastianoF/bruker2nifti#24 @rougier
How to contribute
You should have a CONTRIBUTE file explaining how to contribute to the software (testing, reporting issue, contributing code), etc.
