### 2017-08-14 SebastianoF/bruker2nifti#20 @SebastianoF
Really agree this part requires some improvements. Also there were too many redundancies between the wiki and the README.
I tried to make it more clear while removing redundancies.
