### 2017-08-14 3654a6e
DOC typos
