### 2017-08-14 38b8d6b
DOC typos
