### 2017-08-14 SebastianoF/bruker2nifti#20 @rougier
Also, do you need the virtual environment at all ? I tested it using the git clone/ install -e and it works fine. I found the virtual environment a bit confusing since you do not really explain how to use it (apart from the given link).
