### 2017-08-14 498c1a4
DOC doi added for one out of two, .tex and .pdf removed
