+++ paper/paper.bib
@@ -1,19 +1,27 @@
 @article{cox2004sort,
-	title={A (sort of) new image data format standard: Nifti-1},
-	author={Cox, Robert W and Ashburner, John and Breman, Hester and Fissell, Kate and Haselgrove, Christian and Holmes, Colin J and Lancaster, Jack L and Rex, David E and Smith, Stephen M and Woodward, Jeffrey B and others},
-	journal={Neuroimage},
-	volume={22},
-	pages={e1440},
-	year={2004}
+	title   = {A (sort of) new image data format standard: Nifti-1},
+	author  = {{Cox, Robert W and Ashburner, John and Breman, Hester and Fissell, Kate and Haselgrove, Christian and Holmes, Colin J and Lancaster, Jack L and Rex, David E and Smith, Stephen M and Woodward, Jeffrey B and others}},
+	journal = {Neuroimage},
+	volume  = {22},
+	pages   = {e1440},
+	year    = {2004},
+	ee      = {https://nifti.nimh.nih.gov/nifti-1/documentation/hbm_nifti_2004.pdf},
+	
 }
 
-@inproceedings{mustra2008overview,
-	title={Overview of the DICOM standard},
-	author={Mustra, Mario and Delac, Kresimir and Grgic, Mislav},
-	booktitle={ELMAR, 2008. 50th International Symposium},
-	volume={1},
-	pages={39--44},
-	year={2008},
-	organization={IEEE}
+@article{Mildenberger2002,
+	author  = {Mildenberger, Peter and Eichelberg, Marco and Martin, Eric},
+	title   = {Introduction to the DICOM standard},
+	journal = {European Radiology},
+	year    = {2002},
+	month   = {Apr},
+	day     = {01},
+	volume  = {12},
+	number  = {4},
+	pages   = {920--927},
+	issn    = {1432-1084},
+	doi     = {10.1007/s003300101100},
+	url     = {https://doi.org/10.1007/s003300101100}
 }
+
 
