+++ paper/paper.tex
@@ -45,7 +45,7 @@ The  \href{https://www.bruker.com/products/mr/preclinical-mri/software/service-s
 %A possible solution to address these time consuming and possibly frustrating issues is to avoid the passage trough 2 black boxes from Native format to DICOM and subsequently from DICOM to nifti. This can be done gaining control directly over the Native format, and parsing the native variables that are relevant to fill the nifti header and the consequent image volume.
 
 For this purpose we have designed and developed \href{https://github.com/SebastianoF/bruker2nifti}{\emph{bruker2nifti}}, a pip-installable Python tool provided with a Graphical User Interface to convert from the native MRI Bruker format %\cite{brukerPV6Webiste}, \cite{brukerPreclinicalMRI} 
-to the nifti format, without any intermediate step through the DICOM standard formats \cite{mustra2008overview}. %A command line interface is provided as well to allow for the automatic conversion of a dataset within a pipeline. 
+to the nifti format, without any intermediate step through the DICOM standard formats \cite{Mildenberger2002}. %A command line interface is provided as well to allow for the automatic conversion of a dataset within a pipeline. 
 
 %Finally, as data conversions are never free of deceptions (different scanner settings, different Bruker Paravision software versions and different hardware, can give raise to different parsing and native format structure), 
 Bruker2nifti is intended to be a tool to access the data structure and to parse every Parameter Files of the Bruker ParaVision format into python dictionaries, to select the relevant information to fill the Nifti header and data volume. Lastly it is meant to be a starting point where to integrate possible future variations in Bruker hardware and ParaVision software future releases.
