+++ paper/paper.tex
@@ -1,63 +1,0 @@
-\documentclass{article}
-\usepackage{authblk}
-%\usepackage{cite}
-%\usepackage[square,sort,comma]{natbib}
-\usepackage{hyperref}
-
-\title{Bruker2nifti: Magnetic Resonance Images converter from Bruker ParaVision to Nifti format}
-
-\author[1]{Sebastiano Ferraris}
-\author[1]{Dzhoshkun Ismail Shakir}
-\author[2]{Johannes Van Der Merwe}
-\author[3]{Willy Gsell}
-\author[1,2,4]{Jan Deprest}
-\author[1,2,4]{Tom Vercauteren}
-
-\affil[1]{Translational Imaging Group, Centre for Medical Image Computing (CMIC), Department of Medical Physics and Bioengineering, University College London, UK}
-\affil[2]{Department of Development and Regeneration, Organ System Cluster, Group Biomedical Sciences, KU Leuven, Belgium.}
-\affil[3]{Biomedical MRI, Department of Imaging and Pathology, KU Leuven, Belgium.}
-\affil[4]{Wellcome/EPSRC Centre for Interventional and Surgical Sciences, University College London, UK.}
-
-%\date{}                     %% if you don't need date to appear
-\setcounter{Maxaffil}{0}
-
-\renewcommand\Affilfont{\itshape\small}
-
-\begin{document}
-	\maketitle
-
-%\section{Summary}
-%
-%In this work we present the pip-installable python library \emph{bruker2nifti}. It consists of a medical image format converter from Bruker MRI (ParaVision) to Nifti format. It is entirely written in python, providing a parser tool for the Bruker data structure, a converter to the nifti format (nifti-1 or nifti-2), a graphical user interface and a command line utility.
-
-\section{Motivations and Summary}
-
-%In medical imaging, after a Magnetic Resonance Image dataset is acquired, the first step of many clinical and pre-clinical study is the image conversion from the native format to a format suitable for the intended analysis.
-%Not all proprietary software provide the tools for the data conversion to a suitable open formats for research, such as nifti \cite{cox2004sort} for which most of the available tools for medical image analysis are implemented. 
-
-In clinical and pre-clinical research involving medical images, the first step following a Magnetic Resonance Imaging dataset acquisition, usually entails the image conversion from the native scanner format to a format suitable for the intended analysis. 
-The  \href{https://www.bruker.com/products/mr/preclinical-mri/software/service-support.html}{Bruker ParaVision} proprietary software is not currently providing the tools for the data conversion to a suitable open formats for research, such as nifti \cite{cox2004sort}, for which most of the available tools for medical image analysis are implemented. 
-
-
-%The DICOM (Digital Imaging and Communications in Medicine) standard \cite{mustra2008overview}, was initially designed to provide a common standard output, and to solve this issue. Over the years the standard was bent in different directions to cope with the growing needs of each producer: in the present each scanner manufacturer can have its own DICOM version, so that a DICOM from e.g. a Philips scanner can have a different file structure, file extension and orientation convention, when compared to a DICOM from a Bruker scanner (ParaVision software).
-%Moreover, once the DICOM is obtained from the native format, this requires to be converted again to nifti using one of the available converter that may be well tested from a DICOM obtained from a specific producer, but may introduce other errors, e.g. flipping the R-L orientation or stacking the slices of 4D volumes in the wrong order when applied to DICOM from a different producer.
-
-%A possible solution to address these time consuming and possibly frustrating issues is to avoid the passage trough 2 black boxes from Native format to DICOM and subsequently from DICOM to nifti. This can be done gaining control directly over the Native format, and parsing the native variables that are relevant to fill the nifti header and the consequent image volume.
-
[…]
