### 2017-08-14 SebastianoF/bruker2nifti#23 @SebastianoF
Typo corrected, thanks again.
The paper cox2004sort has no DOI, or no DOI I could find via the link https://www.crossref.org/guestquery/
I added the web link. Also the second reference had no DOI so I changed to an equivalent publication provided with a DOI.
