### 2017-08-14 SebastianoF/bruker2nifti#23 @rougier
New PDF is missing from the repo.
