### 2017-08-14 6faa0a9
DOC README virtualenv as suggestion
