### 2017-08-14 SebastianoF/bruker2nifti#20 @SebastianoF
Uhm, lets say that are warmly suggested. Going into details of the virtualenv can be a bit of an overkill. And could not do it better than The Hitchhiker’s Guide to Python linked.
It is now a parenthesis for the curious reader. Let me know if it makes more sense.
