### 2017-08-14 SebastianoF/bruker2nifti#23 @SebastianoF
About the pdf, wanted to ask if there is an automatic way of creating it from the .md.

I saw in many others JOSS submitted repo that the pdf is not present in the repo, but there is a link that provides a pdf obtained in some sort of standard template.

e.g.
[http://joss.theoj.org/papers/e40b3701ae4f96e1da1caedddafa121a](http://joss.theoj.org/papers/e40b3701ae4f96e1da1caedddafa121a)
[https://github.com/dhermes/bezier](https://github.com/dhermes/bezier)

Do I have to provide the pdf as well, inside the repo? Sorry this is my first JOSS submission.
