### 2017-08-14 SebastianoF/bruker2nifti#23 @rougier
And this is my first review... I don't know actually. We can keep it as it is, I can read the .md.
