### 2017-08-14 ab65717
DOC contributing added, BUG commented the versionig using git number due to Travis-CI clash
