+++ bruker2nifti/__init__.py
@@ -11,9 +11,9 @@ here = os.path.abspath(os.path.dirname(__file__))
 git_dir = os.path.dirname(here)
 
 # Describe the version relative to last tag
-command_git = ['git', 'describe', '--match', 'v[0-9]*']
-version_buf = check_output(command_git, cwd=git_dir).rstrip()
+# command_git = ['git', 'describe', '--match', 'v[0-9]*']
+# version_buf = check_output(command_git, cwd=git_dir).rstrip()
 
 # Exclude the 'v' for PEP440 conformity, see
 # https://www.python.org/dev/peps/pep-0440/#public-version-identifiers
-__version__ = version_buf[1:].decode("utf-8")
+__version__ = '1.0.0'  # version_buf[1:].decode("utf-8")
