### 2017-08-14 929107a
DOC contributing typos
