+++ CONTRIBUTE.md
@@ -1,0 +1,67 @@
+## Contributing to bruker2nifti
+
+Thanks for contributing to bruker2nifti!
+
+In any study using MR Bruker scanners, there are multiple scanner settings, 
+ParaVision versions and MRI protocol. Therefore an open source collaborative environment is the best 
+place where to build an optimal converter that takes into account as many variations as possible.
+
+
+#### Table Of Contents
+
+[Code of Conduct](#code-of-conduct)
+
+[Questions and Issues](#questions-and-issues)
+
+[What should I know?](#what-you-should-know)
+
+[How to contribute?](#how-to-contribute)
+
+[Styleguides](#styleguides)
+
+
+
+### Code of Conduct
+
+This project adopts the [Covenant Code of Conduct](https://contributor-covenant.org/). 
+By participating, you are expected to uphold this code. 
+
+
+### Questions and issues 
+
+For any issue or question related to the code, please raise an issue in the 
+[bruker2nifti issue page](https://github.com/SebastianoF/bruker2nifti/issues).
+Here is where all the comunications between contributors will happen.
+ 
+ 
+### What you should know 
+
+You should be familiar with the bruker ParaVision folder structure and the NifTi (1 and 2)
+format.
+A starting point can be the links in the [README](https://github.com/SebastianoF/bruker2nifti/blob/master/README.md) 
+and the [wiki page](https://github.com/SebastianoF/bruker2nifti/wiki).
+
+The bruker2nifti code structure and API is as simple as possible, see in particular this 
+[wiki-page](https://github.com/SebastianoF/bruker2nifti/wiki/Code-rationale,-definitions-and-structure). 
+
+### How to contribute
+
+To contribute in a constructive way you should try to follow the 
+[proposed code structure and nomenclature](https://github.com/SebastianoF/bruker2nifti/wiki/Code-rationale,-definitions-and-structure). unless there
+are not good reasons to not do it.
+
+
+You are welcome to report bugs and suggesting enhancements in the [bruker2nifti issue page](https://github.com/SebastianoF/bruker2nifti/issues).
+Please use a new issue for each thread.
+
+If you forked the repository and made some contribution that you would like to share, you can do a 
+[git pull request](https://yangsu.github.io/pull-request-tutorial/). Please take a look a the Styleguide below as well.
+
+
+### Styleguides
+
+The code follows the [PEP-8](https://www.python.org/dev/peps/pep-0008/) style convention. The commit names, when possible
+ should follow the [ITK standard prefix commit message convention](https://itk.org/Wiki/ITK/Git/Develop). 
+
+
+
