### 2017-08-14 6b0012c
DOC contributing table of content removed
