### 2017-08-14 2aa253d
DOC contributing typos
