### 2017-08-14 SebastianoF/bruker2nifti#24 @SebastianoF
Initial CONTRIBUTE file added.
