### 2017-08-14 SebastianoF/bruker2nifti#22 @SebastianoF
Uhm... The content of the docs folder is created by sphinx automatically to produce the API documentation
[http://bruker2nifti.readthedocs.io/en/latest/](http://bruker2nifti.readthedocs.io/en/latest/)
I am not sure if the folder docs has to be added to .gitignore or not, as it can be distracting for the

About the real-data, I'm updating the wiki page with some examples!
