### 2017-08-14 92e0054
ENH help and nifti version GUI added
