### 2017-08-14 0999f3f
ENH check if the study folder is a study or if it may be something else
