### 2017-08-14 SebastianoF/bruker2nifti#21 @SebastianoF
The GUI is raw rather for limits in using tkinter.
I added the help button redirecting to the documentation in the development version.
You can check it out installing in development mode. Will be included in the next release!
