### 2017-08-14 SebastianoF/bruker2nifti#22 @rougier
Still, you can modify the index.rst file since it is the master file (from you conf.py). Not much is needed but the current content is rather minimal.
