### 2017-08-14 #354 @Kevin-Mattheus-Moerman
Thanks @rougier!
