### 2017-08-14 46f40b5
BUG parsers in refreshing variable lists, corrected
