### 2017-08-14 b66f3f6
DOC index.rst after issue #22
