### 2017-08-14 SebastianoF/bruker2nifti#22 @SebastianoF
Done! I'm not sure if it will be overwritten in re-creating the documentation with Sphynx. Will try and find another solution to make this part of the documentation more talkative!
