### 2017-08-14 SebastianoF/bruker2nifti#22 @rougier
I think the doc is in rst format and you used markdown for links. You need to modify them as:

[text](link) -> `text <link>`_

Apart from that, I think you're good. I would personally extend the index a little bit but it's up to you at this point.
