### 2017-08-14 #354 @rougier
I finished the review and @SebastianoF made all the changes, I think it can be accepted and is now ready for publication. See:

* https://github.com/SebastianoF/bruker2nifti/issues/17
* https://github.com/SebastianoF/bruker2nifti/issues/18
* https://github.com/SebastianoF/bruker2nifti/issues/20
* https://github.com/SebastianoF/bruker2nifti/issues/21
* https://github.com/SebastianoF/bruker2nifti/issues/22
* https://github.com/SebastianoF/bruker2nifti/issues/23
* https://github.com/SebastianoF/bruker2nifti/issues/24

One remaining question is the possibility to update the submission with version 1.0.1 that is Python 3 compatible. I think it would be better to publish this version (which is also back compatible with Python 2.7).

Last very minor point is the PDF output missing from the source repo but JOSS will recompile it anyway, no?
