### 2017-08-15 c2346fc
DOC index.rst links correction -issue #22
