### 2017-08-15 8544e8f
DOC massive refactoring README
