+++ README.md
@@ -1,122 +1,62 @@
 [![status](http://joss.theoj.org/papers/2ee6a3a3b1a4d8df1633f601bf2b0ffe/status.svg)](http://joss.theoj.org/papers/2ee6a3a3b1a4d8df1633f601bf2b0ffe)
 [![Build Status](https://travis-ci.org/SebastianoF/bruker2nifti.svg?branch=master)](https://travis-ci.org/SebastianoF/bruker2nifti)
 
-## Bruker2nifti
+# Bruker2nifti
 
-Medical image format converter: from raw [Bruker](http://imaging.mrc-cbu.cam.ac.uk/imaging/FormatBruker) 
-ParaVision to [NifTi](https://nifti.nimh.nih.gov/nifti-1), written in Python.
+[Bruker2nifti](https://github.com/SebastianoF/bruker2nifti) is an open source medical image format converter from raw [Bruker](http://imaging.mrc-cbu.cam.ac.uk/imaging/FormatBruker) 
+ParaVision to [NifTi](https://nifti.nimh.nih.gov/nifti-1), without any intermediate step through the DICOM standard formats.
 
-Further information about the motivation can be found [here](https://github.com/SebastianoF/bruker2nifti/tree/master/paper).
-Code repository can be found [here](https://github.com/SebastianoF/bruker2nifti).
+Bruker2nifti is a pip-installable Python tool provided with a Graphical User Interface and a Command Line Utility.
 
-### Requirements
-* Python 2.7+ and the libraries listed in the file [requirements.txt](https://github.com/SebastianoF/bruker2nifti/blob/master/requirements.txt).
+Please note that the stable release is compatible only with **Python 2**. The development release is Python 2 and Python 3 compatible.
 
-### Installation
-To install the latest (stable) release as a python library via pip, type 
-```
-pip install bruker2nifti
-```
-(installing the library inside a [virtualenvironment](http://docs.python-guide.org/en/latest/dev/virtualenvs/) is warmly recommended).
+## Getting Started
++ Requirements
+    - Python 2.7+ 
+    - Libraries in [requirements.txt](https://github.com/SebastianoF/bruker2nifti/blob/master/requirements.txt).
 
-To install the current development version, please see the 
-[wiki-installing](https://github.com/SebastianoF/bruker2nifti/wiki/Installing-latest-updates-and-in-development-mode) page. 
++ Installation
+    - Install the latest stable release with `pip install bruker2nifti`.
+    - [Install the latest development version](https://github.com/SebastianoF/bruker2nifti/wiki/Installing-stable-version-and-development-version).
 
-### Usage:
++ Real data examples
+    - [Python command shell](https://github.com/SebastianoF/bruker2nifti/wiki/Example:-use-bruker2nifti-via-Python-command-shell).
+    - [Command Line Utility (CLI)](https://github.com/SebastianoF/bruker2nifti/wiki/Example:-use-bruker2nifti-via-Command-Line-Utility).
+    - [Graphical User Interface (GUI)](https://github.com/SebastianoF/bruker2nifti/wiki/Graphical-User-Interface-Examples).
 
-The converter can be accessed via: 
-+ Python command shell, 
-+ Command Line Utility (CLI) to integrate it in a bash script,
-+ Graphical User Interface (GUI).
-
-Moreover, a launcher can be automatically created to access the GUI with no-Python knowledge required.  
-
-#### Python command shell
-After installing the stable release with `pip install bruker2nifti`, access the Python command shell typing `python` (or `ipython` if the 
-[interactive command](https://en.wikipedia.org/wiki/IPython) shell is preferred), then import bruker2nifti with
-```
-import bruker2nifti
-```
[…]
