### 2017-08-15 e79f2f3
DOC typos + gui example
