### 2017-08-15 6ecd960
DOC Acknowledgements updates and links
