### 2017-08-15 SebastianoF/bruker2nifti#20 @SebastianoF
@rougier: Hi,
Due to the redundancies between the wiki and the README I did some documentation refactoring.
The informations are the same, probably more accessible and less repetitive. If you have a spare minute let me know if you prefer it now.
