### 2017-08-15 f8464fa
DOC typos
