### 2017-08-15 SebastianoF/bruker2nifti#20 @rougier
Yep, definitely nicer. By the way, I recommended you submission for publication (I didn't manage to mention you using the @ notation). Only remaining issue being the new version to replace the old one and the automatic building of the PDF (very minor issues).
