### 2017-08-15 SebastianoF/bruker2nifti#20 @SebastianoF
@rougier Thanks! [Saw it](https://github.com/openjournals/joss-reviews/issues/354)! There is still the Functionality documentation tick to tick.
Very curious to read the answers about the PDF building and the versioning issue.
