### 2017-08-15 255a33d
DOC typos again
