### 2017-08-15 90a4845
DOC README update + index.rst update
