### 2017-08-15 d950a83
DOC CONTRIBUTE update
