+++ CONTRIBUTE.md
@@ -1,40 +1,16 @@
-## Contributing to bruker2nifti
+# Contributing to bruker2nifti
 
 Thanks for contributing to bruker2nifti!
 
 In any study using MR Bruker scanners, there are multiple scanner settings, 
-ParaVision versions and MRI protocols involved. An open source collaborative environment is a good place
- to where to build a converter that takes into account as many settings as possible.
+ParaVision versions and MRI protocols involved. Take this into consideration when proposing a new contribution.
 
-<!--
-#### Table Of Contents
-
-[Code of Conduct](#code-of-conduct)
-
-[Questions and Issues](#questions-and-issues)
-
-[What should I know?](#what-you-should-know)
-
-[How to contribute?](#how-to-contribute)
-
-[Styleguides](#styleguides)
--->
-
-
-### Code of Conduct
+## Code of Conduct
 
 This project adopts the [Covenant Code of Conduct](https://contributor-covenant.org/). 
 By participating, you are expected to uphold this code. 
-
-
-### Questions and issues 
-
-For any issue or question related to the code, please raise an issue in the 
-[bruker2nifti issue page](https://github.com/SebastianoF/bruker2nifti/issues).
-Here is where all the comunications between contributors will happen.
  
- 
-### What you should know 
+## What you should know 
 
 You should be familiar with the bruker ParaVision folder structure and the NifTi (1 and 2)
 format.
@@ -44,24 +20,23 @@ and the [wiki page](https://github.com/SebastianoF/bruker2nifti/wiki).
 The bruker2nifti code structure and API is as simple as possible, see in particular this 
 [wiki-page](https://github.com/SebastianoF/bruker2nifti/wiki/Code-rationale,-definitions-and-structure). 
 
-### How to contribute
+## Contributions: Questions, bugs, issues and new features 
 
-To contribute in a constructive way you should try to follow the 
-[proposed code structure and nomenclature](https://github.com/SebastianoF/bruker2nifti/wiki/Code-rationale,-definitions-and-structure). unless there
-are not good reasons to not do it.
++ For any issue bugs or question related to the code, please raise an issue in the 
+[bruker2nifti issue page](https://github.com/SebastianoF/bruker2nifti/issues).
 
++ Propose here as well improvements suggestions and new features.
 
-You are welcome to report bugs and suggesting enhancements in the [bruker2nifti issue page](https://github.com/SebastianoF/bruker2nifti/issues).
-Please use a new issue for each thread.
++ **Please use a new issue for each thread.** Issue re-usable by other users that may have a similar problem.
 
-If you forked the repository and made some contribution that you would like to share, you can do a 
-[git pull request](https://yangsu.github.io/pull-request-tutorial/). Please take a look a the Styleguide below as well.
++ To contribute in a coherent way please follow the 
+[proposed code structure and nomenclature](https://github.com/SebastianoF/bruker2nifti/wiki/Code-rationale,-definitions-and-structure), .
 
++ If you forked the repository and made some contributions that you would like to integrate in the git master branch, you can do a 
+[git pull request](https://yangsu.github.io/pull-request-tutorial/). 
+ 
 
-### Styleguides
+## Styleguides
 
-The code follows the [PEP-8](https://www.python.org/dev/peps/pep-0008/) style convention. The commit names, when possible
- should follow the [ITK standard prefix commit message convention](https://itk.org/Wiki/ITK/Git/Develop). 
-
-
-
++ The code follows the [PEP-8](https://www.python.org/dev/peps/pep-0008/) style convention. 
++ Please follow the [ITK standard prefix commit message convention](https://itk.org/Wiki/ITK/Git/Develop) for commit messages. 
