### 2017-08-16 6e21bd7
ENH extra method apply_orientation_matrix_to_image
