### 2017-08-16 bcbfdaf
ENH extra apply_orient correct + precision modified
