### 2017-08-16 01ebfa6
DOC index .rts issues in bulletpoints visualisation
