### 2017-08-16 9cd91ae
BUG issue #28 corrected
