### 2017-08-16 842a430
BUG issue #28 correction extended for bruker2nifti_scan
