### 2017-08-16 f8e5702
DOC typos
