+++ bruker2nifti/parsers/bruker2nii_scan.py
@@ -7,7 +7,8 @@ from bruker2nifti.converter import Bruker2Nifti
 def main_scan():
     """
     Parser from terminal with
-    $ python2 bruker2nifti_study -i input_file_path -o output_file_path
+    $ python2 bruker2nifti_scan -h
+    $ python2 bruker2nifti_scan -i input_file_path -o output_file_path
     """
 
     parser = argparse.ArgumentParser()
