+++ LICENCE.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2017 Sebastiano Ferraris
+Copyright (c) 2017 Sebastiano Ferraris (University College London)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
