### 2017-08-19 d97d0e3
bug fix: slope corrected vol_data was being replaced by original data
