### 2017-08-19 01c95c5
add rename slope correction to visu_slope correction, and add reco_slope correction
