+++ bruker2nifti/_cores.py
@@ -12,7 +12,8 @@ from ._utils import bruker_read_files, normalise_b_vect, from_dict_to_txt_sorted
 
 
 def scan2struct(pfo_scan,
-                correct_slope=True,
+                correct_visu_slope=True,
+                correct_reco_slope=True,
                 nifti_version=1,
                 qform_code=1,
                 sform_code=2,
@@ -33,7 +34,8 @@ def scan2struct(pfo_scan,
     Data are parsed in the intermediate dictionary struct containing the final scan(s) converted in nibabel
     image, with additional infos.
     :param pfo_scan: path to folder containing the scan
-    :param correct_slope: [True] if you want to correct the slope of the values.
+    :param correct_visu_slope: [True] if you want to correct the visu_slope of the values.
+    :param correct_reco_slope: [True] if you want to correct the reco_slope of the values.
     :param nifti_version: [1] output nifti version can be version 1 or version 2 (see nibabel documentation)
     :param qform_code: [1] qform of the final nifti image
     :param sform_code: [2] sform of the final nifti image
@@ -71,8 +73,17 @@ def scan2struct(pfo_scan,
 
     for id_sub_scan in list_sub_scans:
 
-        visu_pars = bruker_read_files('visu_pars', pfo_scan, sub_scan_num=id_sub_scan)
+        print "Processing scan", id_sub_scan
+        
+        reco = {}
+        reco = bruker_read_files('reco', pfo_scan, sub_scan_num=id_sub_scan)
+        if reco == {}:
+            warn_msg = "Warning: No 'reco' file to parse in scan "+id_sub_scan
+            warnings.warn(warn_msg)
+            correct_reco_slope = False
 
+        visu_pars = {}
+        visu_pars = bruker_read_files('visu_pars', pfo_scan, sub_scan_num=id_sub_scan)
         if visu_pars == {}:
             warn_msg = "\nNo 'visu_pars' data found here: \n{}. \nAre you sure the input folder contains a " \
                        "proper Bruker scan?\n".format(jph(pfo_scan, 'pdata', id_sub_scan))
@@ -119,18 +130,24 @@ def scan2struct(pfo_scan,
         is_dwi = 'dtiepi' in visu_pars_acq_sequence_name.lower()
 
         if is_dwi:
-            # Force to not correcting the slope, if true. Diffusion weighted images must be slope corrected before the
+            # Force to not correcting the visu_slope, if true. Diffusion weighted images must be slope corrected before the
             # DTI analysis. They will be to heavy otherwise.
-            correct_slope = False
-            # Force method to be parsed. Useful infos in this file to process the DWI.
+            # (for reco_slope, the user decision is not changed)
+            correct_visu_slope = False
+
+            # Force method file to be parsed. Useful infos in this file to process the DWI.
             get_method = True
+            # Force reco file to be parsed. Useful infos in this file to process the DWI.
+            get_reco   = True
 
         # ------------------------------------------------------ #
         # ------ Generate the nifti image using visu_pars. ----- #
         # ------------------------------------------------------ #
         nib_im = nifti_getter(img_data_vol,
                               visu_pars,
-                              correct_slope,
+                              reco,
+                              correct_visu_slope,
+                              correct_reco_slope,
                               nifti_version,
                               qform_code,
                               sform_code,
@@ -301,8 +318,8 @@ def write_struct(bruker_struct,
         # A) Save visu_pars for each sub-scan:
[…]
