### 2017-08-19 ed28fe1
add apply_matrix switch to apply a user-provided 4x4 affine transformation matrix (stored in .txt format)
