### 2017-08-19 3e8b703
STYLE unused apply_orientation_matrix_to_image removed
