### 2017-08-19 0986efe
DOC CONTRIBUTE, improved contributions guidelines added after contribution #30.
