### 2017-08-19 1dec79e
DOC CONTRIBUTE more guidelines
