### 2017-08-21 #354 @Kevin-Mattheus-Moerman
Thanks @rougier. Are we good to tick the last box? @arfon can you answer the PDF query

> Last very minor point is the PDF output missing from the source repo but JOSS will recompile it anyway, no?
