### 2017-08-21 #354 @Kevin-Mattheus-Moerman
@SebastianoF I've some minor editorial comments on your paper below. Feel free to disagree.

I suggest you change
> usually entails the image conversion from the native

to:
> usually entails the conversion of image data from the native

I suggest you change:
> The Bruker ParaVision proprietary software is not currently providing the tools for the data conversion to a suitable open formats for research, such as nifti [@cox2004sort], for which most of the available tools for medical image analysis are implemented.

to:
> The proprietary Bruker ParaVision software currently does not provide the tools for conversion of the data to suitable and open formats for research, such as nifti [@cox2004sort], for which most of the available tools for medical image analysis are implemented.

In the following the sentence does not read well and and use of capital letters seems odd:
> ... and to parse every Parameter Files of the Bruker ParaVision format into python dictionaries...

Should this instead become?:
> ... and to parse all parameter files for the Bruker ParaVision format into python dictionaries...

Could the following sentence:
> Lastly it is meant to be a starting point where to integrate possible future variations in Bruker hardware and ParaVision software future releases.

Be rephrased to something like:
> Lastly it is meant to be a starting point to integrate possible future variations in Bruker hardware and ParaVision software.
