### 2017-08-21 28418f4
DOC paper.md corrections after JOSS review openjournals/joss-reviews#354
