+++ paper/paper.md
@@ -40,13 +40,13 @@ bibliography: paper.bib
 
 # Motivations and Summary
 
-In clinical and pre-clinical research involving medical images, the first step following a Magnetic Resonance Imaging dataset acquisition, usually entails the image conversion from the native scanner format to a format suitable for the intended analysis. 
-The [Bruker ParaVision](https://www.bruker.com/products/mr/preclinical-mri/software/service-support.html) proprietary software is not currently providing the tools for the data conversion to a suitable open formats for research, such as nifti [@cox2004sort], for which most of the available tools for medical image analysis are implemented. 
+In clinical and pre-clinical research involving medical images, the first step following a Magnetic Resonance Imaging dataset acquisition, usually entails the conversion of image data from the native scanner format to a format suitable for the intended analysis. 
+The proprietary [Bruker ParaVision](https://www.bruker.com/products/mr/preclinical-mri/software/service-support.html) software currently does not provide the tools for conversion of the data to suitable and open formats for research, such as nifti [@cox2004sort], for which most of the available tools for medical image analysis are implemented. 
 
 For this purpose we have designed and developed [bruker2nifti](https://github.com/SebastianoF/bruker2nifti), a pip-installable Python tool provided with a Graphical User Interface to convert from the native MRI Bruker format 
 to the nifti format, without any intermediate step through the DICOM standard formats [@Mildenberger2002].
 
-Bruker2nifti is intended to be a tool to access the data structure and to parse every Parameter Files of the Bruker ParaVision format into python dictionaries, to select the relevant information to fill the Nifti header and data volume. Lastly it is meant to be a starting point where to integrate possible future variations in Bruker hardware and ParaVision software future releases.
+Bruker2nifti is intended to be a tool to access the data structure and to parse all parameter files of the Bruker ParaVision format into python dictionaries, to select the relevant information to fill the Nifti header and data volume. Lastly it is meant to be a starting point where to integrate possible future variations in Bruker hardware and ParaVision software future releases.
 
 # Acknowledgements
 
