### 2017-08-22 #354 @arfon
> Last very minor point is the PDF output missing from the source repo but JOSS will recompile it anyway, no?

Yep, I'll compile that shortly.
