### 2017-08-22 #354 @SebastianoF
@arfon Hi, .zip uploaded to [https://zenodo.org/record/846814](https://zenodo.org/record/846814).
The DOI is [10.5281/zenodo.846814](https://zenodo.org/badge/DOI/10.5281/zenodo.846814.svg)
Thanks!
