### 2017-08-26 6662cde
ENH new get_stack_direction_from_VisuCorePosition method. WIP integrate to interpret the final orientation
