### 2017-08-26 SebastianoF/bruker2nifti#22 @SebastianoF
Closing for now.
Feel free to re-open for further suggestions in this thread. Thanks
